requirements.txt
gym>=0.26.2
opencv-python==4.7.0.68
plotly==5.13.0
torch==2.0.0
torchrl==0.1.1
tqdm==4.64.1
sympy==1.11.1
anytree==2.8.0
torchvision==0.15.1
tensorboard==2.11.2
pyglet==1.5.27
minigrid==2.2.0
pyyaml
matplotlib
pandas